SudokuAnalize

Description

Collection of functions that check specific properties of a sudoku grid or its' subcomponents

Documentation

empty

Arguments

:: Value
Value of a field

-> Bool
True if no concrete value

Check if sudoku field is empty

single

Arguments

:: [a]
usually 'SudokuBase#Choices'

-> Bool
True if it only contains one choice

Check if an array contains only ONE component

valid

Arguments

:: Grid
Sudoku grid that doesn't have to be valid

-> Bool
True if no dublicate entries in each row, column and box

Check if sudoku grid is valid

complete

Arguments

:: Matrix Choices
representing a possibly unsolved sudoku grid

-> Bool
True if every field has a single choice

A Matrix is complete if it only contains single choices

blocked

Arguments

:: Matrix Choices
a possibly unsolvable sudoku grid

-> Bool
True if not solvable

Either void or unsafe