SudokuBase

Description

Basic types and functions needed to represent a sudoku grid

Types used to represent both solved and unsolved sudoku grids

type Value

Arguments

= Char
Value of a field in the sudoku grid

type Row a

Arguments

= [a]
A list that represents a row in a Matrix

type Grid

Arguments

= Matrix Value
Matrix that represents a sudoku grid

type Matrix a

Arguments

= [Row a]
2D array

type Choices

Arguments

= [Value]
List of possible values a field could have

Symetric transformations of a sudoku grid

rows

Arguments

:: Matrix a
Matrix (usually Grid) which is already a list of Rows

-> [Row a]
list of rows in a sudoku grid

Create list of rows

cols

Arguments

:: Matrix a
Matrix (usually Grid)

-> [Row a]
list of rows (using transpose)

Create list of columns

boxs

Arguments

:: Matrix a
Matrix (usually Grid)

-> [Row a]
list of boxes (sectors in a sudoku grid)

Create list of boxes (boxsize x boxsize)

Constants defining a sudoku grid

values :: [Value]

Char values for all allowed numbers

boxsize :: Int

Size of sectors in a sudoku grid