SudokuSolver

Description

This module contains the toplevel function to solve sudoku grids as well as its' sub functions.

Documentation

solve

Arguments

:: Grid
(unsolved) Grid as input

-> [Grid]
list whicht eigher contains a solution or is empty

Tries to find a solution to a specified Grid