SudokuTransform

Description

A collection of functions that return a somehow transformed sudoku grid.

Documentation

choices

Arguments

:: Grid
Grid that contains empty fields

-> Matrix Choices
'Matrix Choices' containing either ['1'..'9'] or a single valued array

replaces empty fields with Choices.

collapse

Arguments

:: Matrix [a]
Matrix of choices

-> [Matrix a]
list of all possible permutations considering the choices

Creates a list of matrices that mustn't be valid representing all permutation of the input (reduced by one dimension).

prune

Arguments

:: Matrix Choices
contains invalid choices

-> Matrix Choices
all invalid choices are removed

Remove all invalid choices by checking rows columns and boxes of the Sudoku grid.